zombos/requirements.txt
pygame==2.5.2
Pillow==10.2.0 